// src/app/core/models/product.interface.ts
import { Category } from './category.interface';

export interface Product {
  sold: number;
  images: string[];
  subcategory: Category[];
  ratingsQuantity: number;
  _id: string;
  title: string;
  slug: string;
  description: string;
  quantity: number;
  price: number;
  priceAfterDiscount?: number;
  imageCover: string;
  category: Category;
  brand: Category;
  ratingsAverage: number;
  createdAt: string;
  updatedAt: string;
  id: string;
}